push http/progress utils up to 8mi
