+++ layers/eight_mile/progress.py
@@ -1,11 +1,18 @@
 import re
 import sys
 import six
-from eight_mile.utils import exporter
+from eight_mile.utils import exporter, register, optional_params
 
 
 __all__ = []
 export = exporter(__all__)
+MEAD_LAYERS_PROGRESS = {}
+
+
+@export
+@optional_params
+def register_progress(cls, name=None):
+    return register(cls, MEAD_LAYERS_PROGRESS, name, 'progress')
 
 
 @export
@@ -44,6 +51,7 @@ class Progress(object):
 
 
 @export
+@register_progress('jupyter')
 class ProgressBarJupyter(Progress):
     """Simple Jupyter progress bar
 
@@ -71,6 +79,7 @@ class ProgressBarJupyter(Progress):
 # Modifed from here
 # http://stackoverflow.com/questions/3160699/python-progress-bar#3160819
 @export
+@register_progress('default')
 class ProgressBarTerminal(Progress):
     """Simple terminal-based progress bar
 
@@ -124,3 +133,12 @@ class ProgressBarTerminal(Progress):
         self.current = self.total
         self.update(step=0)
         self.print_("")
+
+
+@export
+def create_progress_bar(steps, name='default', **kwargs):
+    return MEAD_LAYERS_PROGRESS[name](steps, **kwargs)
+
+
+
+
